[CAS-335] Move CoroutineCall to client library, mark it as internal API, replace FileUploaderCall usages with it
