+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/internal/InternalStreamChatApi.kt
@@ -1,0 +1,7 @@
+package io.getstream.chat.android.client.internal
+
+@RequiresOptIn(
+    message = "This is internal API for the Stream Chat libraries. Do not depend on this API in your own client code.",
+    level = RequiresOptIn.Level.ERROR
+)
+public annotation class InternalStreamChatApi
