--- stream-chat-android-client/src/main/java/io/getstream/chat/android/client/uploader/FileUploaderCall.kt
@@ -1,25 +1,0 @@
-package io.getstream.chat.android.client.uploader
-
-import io.getstream.chat.android.client.call.Call
-import io.getstream.chat.android.client.utils.Result
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-
-internal class FileUploaderCall<T : Any>(
-    private val coroutineScope: CoroutineScope,
-    private val call: () -> Result<T>
-) : Call<T> {
-
-    override fun cancel() {
-        // Empty
-    }
-
-    override fun execute(): Result<T> = call()
-
-    override fun enqueue(callback: (Result<T>) -> Unit) {
-        coroutineScope.launch {
-            val result = call()
-            callback(result)
-        }
-    }
-}
