+++ stream-chat-android-offline/build.gradle
@@ -69,6 +69,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
         freeCompilerArgs += [
                 '-progressive',
                 '-Xexplicit-api=strict',
+                '-Xopt-in=io.getstream.chat.android.client.internal.InternalStreamChatApi',
         ]
     }
 }
