+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/usecase/HideChannelImpl.kt
@@ -1,8 +1,8 @@
 package io.getstream.chat.android.livedata.usecase
 
 import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
 import io.getstream.chat.android.livedata.ChatDomainImpl
-import io.getstream.chat.android.livedata.utils.CoroutineCall
 import io.getstream.chat.android.livedata.utils.validateCid
 
 public interface HideChannel {
