+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/usecase/LoadNewerMessagesImpl.kt
@@ -1,9 +1,9 @@
 package io.getstream.chat.android.livedata.usecase
 
 import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.livedata.ChatDomainImpl
-import io.getstream.chat.android.livedata.utils.CoroutineCall
 import io.getstream.chat.android.livedata.utils.validateCid
 
 public interface LoadNewerMessages {
