+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/usecase/QueryChannelsImpl.kt
@@ -2,11 +2,11 @@ package io.getstream.chat.android.livedata.usecase
 
 import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
 import io.getstream.chat.android.client.utils.FilterObject
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.livedata.ChatDomainImpl
 import io.getstream.chat.android.livedata.controller.QueryChannelsController
-import io.getstream.chat.android.livedata.utils.CoroutineCall
 import kotlinx.coroutines.launch
 
 public interface QueryChannels {
