+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/usecase/QueryChannelsLoadMoreImpl.kt
@@ -2,10 +2,10 @@ package io.getstream.chat.android.livedata.usecase
 
 import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.utils.FilterObject
 import io.getstream.chat.android.livedata.ChatDomainImpl
-import io.getstream.chat.android.livedata.utils.CoroutineCall
 
 public interface QueryChannelsLoadMore {
     /**
