+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/usecase/SendMessageWithAttachmentsImpl.kt
@@ -2,12 +2,12 @@ package io.getstream.chat.android.livedata.usecase
 
 import android.webkit.MimeTypeMap
 import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.livedata.ChatDomainImpl
 import io.getstream.chat.android.livedata.controller.ChannelControllerImpl
-import io.getstream.chat.android.livedata.utils.CoroutineCall
 import io.getstream.chat.android.livedata.utils.validateCid
 import java.io.File
